Fixed GitHub vulnerabilities and removed package-lock.json from remote repository
Fixed GitHub vulnerabilities using a yarn upgrade and removed package-lock.json from remote repository because it is a bad practice to have a package-lock.json and a yarn.lock file for managing the dependencies of a project. The package-lock.json file was probably added in .gitignore after it has been pushed to the remote repository.